+++ CHANGELOG.md
@@ -48,6 +48,8 @@
 
 * `malformed input exception`: Capture MalformedInputExceptions when thrown by reading files with invalid Ascii characters when unsupported by Java (PR #458).
 
+* `viash project file parsing`: Give a more informative message when the viash project file fails to parse correctly (PR #475).
+
 ## INTERNAL CHANGES
 
 * `NextflowVDSL3Platform`: Renamed to `NextflowPlatform` (PR #469).